add authenticate for route

+++ server/src/controller/PositionController.js
@@ -2,6 +2,8 @@ const positionSevice = require('../service/PositionService')
 const { StatusCodes } = require('http-status-codes');
 exports.addPosition = async (req, res, next) => {
     try {
+         if (!req.isAuthenticated && req.role !=  require("../security/JwtUtils").Option.ADMIN)
+            throw new ErrorHandling(403, "You do not have permission");
         const result = await positionSevice.addPosition(req.body);
         res.status(StatusCodes.CREATED).json({
             success: true,
@@ -13,6 +15,8 @@ exports.addPosition = async (req, res, next) => {
 }
 exports.getPositions = async (req, res, next) => {
     try {
+         if (!req.isAuthenticated && req.role !=  require("../security/JwtUtils").Option.ADMIN)
+            throw new ErrorHandling(403, "You do not have permission");
         const result = await positionSevice.getPositions(req.query);
         res.status(StatusCodes.OK).json({
             success: true,
@@ -24,6 +28,8 @@ exports.getPositions = async (req, res, next) => {
 }
 exports.updatePosition = async (req, res, next) => {
     try {
+         if (!req.isAuthenticated && req.role !=  require("../security/JwtUtils").Option.ADMIN)
+            throw new ErrorHandling(403, "You do not have permission");
         const result = await positionSevice.updatePositionById(req.params, req.body);
         res.status(StatusCodes.CREATED).json({
             success: true,
@@ -35,6 +41,8 @@ exports.updatePosition = async (req, res, next) => {
 }
 exports.deletePosition = async (req, res, next) => {
     try {
+         if (!req.isAuthenticated && req.role !=  require("../security/JwtUtils").Option.ADMIN)
+            throw new ErrorHandling(403, "You do not have permission");
         const result = await positionSevice.deletePositionById(req.params);
         res.status(StatusCodes.OK).json({
             success: true,
@@ -47,6 +55,8 @@ exports.deletePosition = async (req, res, next) => {
 }
 exports.getPositionById = async (req, res, next) => {
     try {
+         if (!req.isAuthenticated && req.role !=  require("../security/JwtUtils").Option.ADMIN)
+                    throw new ErrorHandling(403, "You do not have permission");
         const result = await positionSevice.getPositionById(req.params);
         res.status(StatusCodes.OK).json({
             success: true,
